Se refactoriza la estructura de los paketes para el agregado consulta
