+++ Back/applications/app-service/src/main/java/co/com/sofka/config/UseCasesConfig.java
@@ -1,8 +1,8 @@
 package co.com.sofka.config;
 
 import co.com.sofka.model.consulta.gateways.ConsultaRepository;
-import co.com.sofka.model.mascotapokemon.gateways.MascotaPokemonRepository;
-import co.com.sofka.model.usuario.gateways.UsuarioRepository;
+import co.com.sofka.model.consulta.entity.mascotapokemon.gateways.MascotaPokemonRepository;
+import co.com.sofka.model.consulta.entity.usuario.gateways.UsuarioRepository;
 import co.com.sofka.usecase.createconsulta.CreateConsultaUseCase;
 import co.com.sofka.usecase.createmascota.CreateMascotaUseCase;
 import co.com.sofka.usecase.createusuario.CreateUsuarioUseCase;
