+++ Back/domain/model/src/main/java/co/com/sofka/model/consulta/entity/mascotapokemon/gateways/MascotaPokemonRepository.java
@@ -1,6 +1,6 @@
-package co.com.sofka.model.mascotapokemon.gateways;
+package co.com.sofka.model.consulta.entity.mascotapokemon.gateways;
 
-import co.com.sofka.model.mascotapokemon.MascotaPokemon;
+import co.com.sofka.model.consulta.entity.mascotapokemon.MascotaPokemon;
 
 import java.util.List;
