+++ Back/domain/model/src/main/java/co/com/sofka/model/consulta/values/valueobjectmascota/Raza.java
@@ -1,17 +1,17 @@
-package co.com.sofka.model.mascotapokemon.valueobjectmascota;
+package co.com.sofka.model.consulta.values.valueobjectmascota;
 
 import java.util.Objects;
 
-public class Tipo {
+public class Raza {
     private final String valor;
 
-    public Tipo(String valor) {
-        this.valor = Objects.requireNonNull(valor,"el tipo de pokemon es requerido");
+    public Raza(String valor) {
+        this.valor = Objects.requireNonNull(valor,"la raza es requerida");
         if (this.valor.isEmpty()){
-            throw new IllegalArgumentException("el tipo de pokemon no puede ser vacio");
+            throw new IllegalArgumentException("El campo raza no puede ser vacio");
         }
         if (this.valor.length() < 3 || this.valor.length() >= 50){
-            throw new IllegalArgumentException("el tipo de pokemon debe terner mas de 2 caracteres y menos de 50 caracteres");
+            throw new IllegalArgumentException("La raza debe terner mas de 2 caracteres y menos de 50 caracteres");
         }
     }
 
@@ -23,8 +23,8 @@ public class Tipo {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        Tipo tipo = (Tipo) o;
-        return Objects.equals(valor, tipo.valor);
+        Raza raza = (Raza) o;
+        return Objects.equals(valor, raza.valor);
     }
 
     @Override
