+++ Back/domain/usecase/src/main/java/co/com/sofka/usecase/createusuario/CreateUsuarioUseCase.java
@@ -1,7 +1,7 @@
 package co.com.sofka.usecase.createusuario;
 
-import co.com.sofka.model.usuario.Usuario;
-import co.com.sofka.model.usuario.gateways.UsuarioRepository;
+import co.com.sofka.model.consulta.entity.usuario.Usuario;
+import co.com.sofka.model.consulta.entity.usuario.gateways.UsuarioRepository;
 import lombok.RequiredArgsConstructor;
 
 import java.util.List;
